Change Localizer#localize to always return a function
Better to be consistent. To make it more convenient if the message
is a simple string the function has a toString that also returns
the string. See the JSDoc for more details.

Loading of messages inside Montage for localization
Default locale of localizer to defaultLocalizer.locale on init
Update documentation
Update localizer tests

Update localizer to load package.json with a custom require

Before the localizer would always try to load the messages from
the Montage package. Now it accepts a require function to load
from where the require comes from. By default it uses
{global|window}.require, and so will load from the package that
Montage is used in.

Add failing test for creating non-en Localizer